requirements.txt
paho-mqtt
w1thermsensor
loguru
